Python/AI_Assignment_Writer/requirements.txt
google-auth
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client 
beautifulsoup4 
groq 
markdown2
python-dotenv
